pyproject.toml
[tool.poetry]
name = "ai-workflow-platform"
version = "0.1.0"
description = "AI Autonomous Knowledge & Workflow Platform backend"
authors = ["AI Developer <dev@example.com>"]
readme = "README.md"
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
pydantic-settings = "^2.3.4"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
ruff = "^0.5.6"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["backend"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
